Initial DB Refactoring
